refactor: move pages to it's own plugin
