+++ adex/src/ssr.d.ts
@@ -1,17 +1,3 @@
 export { toStatic } from 'hoofd/preact'
 export { renderToString } from 'preact-render-to-string'
-
-export function normalizeRouteImports<Routes>(
-  obj: Routes,
-  matcher: [RegExp, string]
-): {
-  route: string
-  regex: {
-    pattern: RegExp
-    keys: string[]
-  }
-  routePath: string
-  module: () => Promise<{
-    default: () => any
-  }>
-}[]
+export { parse as pathToRegex } from 'regexparam'
